requirements.txt
pillow==11.2.1
setuptools==80.9.0
wheel==0.45.1
